impl: migrate platform plugin to latest 2.x
IntelliJ Platform Gradle Plugin 2.x is the build system that supersedes the Gradle IntelliJ Plugin 1.x, and this is the version that also comes with the ability to provide unlimited until build support.

This is a major overhaul of the DSL, I've tested these changes locally with IntelliJ Community 2022 (the minimum supported
version), and I've also compared the plugin.xml and the zip content with a previous version to make sure
it is generated correctly.